cambios

+++ RobotManager/src/com/gdc/nms/robot/gui/RobotManagerGui.java
@@ -5,6 +5,7 @@ import java.awt.event.ActionListener;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -38,12 +39,16 @@ import com.gdc.nms.robot.gui.tree.TreeListener;
 import com.gdc.nms.robot.gui.tree.TreeModelElements;
 import com.gdc.nms.robot.util.AppExaminator;
 import com.gdc.nms.robot.util.Constants;
+import com.gdc.nms.robot.util.CreatorRobotManager;
 import com.gdc.nms.robot.util.ValidatorManagement;
 import com.gdc.nms.robot.util.indexer.AppInformation;
 import com.gdc.nms.robot.util.jade.InitPlataform;
 import com.gdc.nms.robot.util.jade.SRMAgentManager;
 import com.gdc.nms.robot.util.jade.StatusAgent;
 import com.gdc.nms.robot.util.registry.CommandExecutor;
+import com.gdc.robothelper.webservice.ClientWebService;
+import com.gdc.robothelper.webservice.robot.CreatorRobotWebService;
+import com.gdc.robothelper.webservice.robot.Webservice;
 import com.sun.jmx.mbeanserver.JmxMBeanServerBuilder;
 import com.sun.xml.internal.bind.v2.schemagen.xmlschema.Appinfo;
 
@@ -226,15 +231,45 @@ public class RobotManagerGui extends JFrame {
 				SwingUtilities.invokeLater(new Runnable() {
 					@Override
 					public void run() {
+						System.out.println("servicio de consulta"+checkWebServicesConsult());
+						System.out.println("consultando servicio de creacion"+checkWebServicesCreator());
+							if(checkWebServicesConsult() && checkWebServicesCreator()){
+								SelectorApp selector = new SelectorApp();
+								selector.setVisible(true);
 
-							SelectorApp selector = new SelectorApp();
-							selector.setVisible(true);
-
+							}else{
+								JOptionPane.showMessageDialog(null, "No es posible conectar con el servidor","Error",JOptionPane.ERROR_MESSAGE);
+							}
+							
 					}
 				});
 
 			}
 		});
+	}
+	
+	
+	private boolean checkWebServicesCreator(){
+		URL webServicesCreator = CreatorRobotWebService.getWebServicesCreator();
+		if(webServicesCreator!=null){
+			System.out.println("consultando webservices"+webServicesCreator);
+			return CreatorRobotWebService.existeConexion(webServicesCreator.toString());
+		}else{
+			webServicesCreator=Webservice.getUrl();
+			System.out.println("consultando webservices"+webServicesCreator);
+			return CreatorRobotWebService.existeConexion(webServicesCreator.toString());
+		}
+	}
+	
+	private boolean checkWebServicesConsult(){
+		URL webServicesConsult = ClientWebService.getWebServicesConsult();
+		if(webServicesConsult!=null){
+			System.out.println("comprobando webservices"+webServicesConsult);
+			return ClientWebService.existeConexion(webServicesConsult.toString());
+		}else{
+			webServicesConsult=Webservice.getUrl();
+			return ClientWebService.existeConexion(webServicesConsult.toString());
+		}
 	}
 	private void updateMenuAction(){
 		updateMenu.addActionListener(new ActionListener() {
@@ -263,8 +298,13 @@ public class RobotManagerGui extends JFrame {
 			public void actionPerformed(ActionEvent e) {
 				SwingUtilities.invokeLater( new Runnable() {
 					public void run() {
-						DeleteRobotPanel deleterPanel=new DeleteRobotPanel();
-						deleterPanel.setVisible(true);
+						if(checkWebServicesCreator()){
+							DeleteRobotPanel deleterPanel=new DeleteRobotPanel();
+							deleterPanel.setVisible(true);							
+						}else{
+							JOptionPane.showMessageDialog(null, "No es posible conectar con el servidor","Error",JOptionPane.ERROR_MESSAGE);
+						}
+						
 						
 					}
 				});
